pyproject.toml
[tool.poetry]
name = "crawler-bot"
version = "0.1.0"
description = "Telegram bot for crawling and analyzing prices"
authors = ["Daniil <kononov.miromanov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^22.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
requests = "^2.32.3"
lxml = "^5.3.1"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"
selenium = "^4.29.0"
ipykernel = "^6.29.5"
webdriver-manager = "^4.0.2"
scrapy = "^2.12.0"
scrapy-selenium = "^0.0.7"
undetected-chromedriver = "^3.5.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
crawler-bot = "bot.py:main"
